Add missing features in the app.
