+++ app/api/api.py
@@ -10,7 +10,7 @@ Client = getClient()
 def getRecentTweets():
    tweets = Client.search_recent_tweets(query='"twitter data" (elon musk)',
    expansions= ['author_id'],
-   tweet_fields = ['created_at','public_metrics','possibly_sensitive','attachments','referenced_tweets'], 
+   tweet_fields = ['created_at','public_metrics','possibly_sensitive','attachments','referenced_tweets','entities'], 
    user_fields = ['id','public_metrics','verified', 'created_at'], max_results = 100 )
 
    return tweets
